Create temp account for anons before saving edit
If the user is an anon and temporary accounts are enabled, a
temporary account should be created for the user. Otherwise,
EditPage will throw a "CannotCreateActorException: Cannot
create an actor for an IP user when temporary accounts are
enabled" exception, as the user from the context is used for
permission checks in EditPage::attemptSave.

Change-Id: Ie41f174906ec4da13a10276c896ef8e777bbc6bf